Add files via upload
Afrique Youth Election Observers (AYEO) is a dynamic, youth-led organization committed to promoting democratic engagement and the integrity of electoral processes across the world. Our mission is to empower young people to actively participate in election observation, ensuring that elections are conducted with the utmost transparency, accountability, and fairness. We strive to create a new generation of election observers who are dedicated to upholding democratic values and contributing to credible, peaceful, and inclusive elections.